Move MechanicalGetConstraintInfoVisitor in its own file
