+++ modules/SofaConstraint/src/SofaConstraint/LCPConstraintSolver.cpp
@@ -31,6 +31,9 @@
 #include <sofa/core/ObjectFactory.h>
 #include <sofa/helper/fwd.h>
 
+#include <sofa/simulation/mechanicalvisitor/MechanicalGetConstraintInfoVisitor.h>
+using sofa::simulation::mechanicalvisitor::MechanicalGetConstraintInfoVisitor;
+
 #include <sofa/simulation/mechanicalvisitor/MechanicalVOpVisitor.h>
 using sofa::simulation::mechanicalvisitor::MechanicalVOpVisitor;
 
@@ -44,56 +47,6 @@ using sofa::core::VecId;
 
 namespace sofa::component::constraintset
 {
-
-class MechanicalGetConstraintInfoVisitor : public simulation::BaseMechanicalVisitor
-{
-public:
-    typedef core::behavior::BaseConstraint::VecConstraintBlockInfo VecConstraintBlockInfo;
-    typedef core::behavior::BaseConstraint::VecPersistentID VecPersistentID;
-    typedef core::behavior::BaseConstraint::VecConstCoord VecConstCoord;
-    typedef core::behavior::BaseConstraint::VecConstDeriv VecConstDeriv;
-    typedef core::behavior::BaseConstraint::VecConstArea VecConstArea;
-
-    MechanicalGetConstraintInfoVisitor(const core::ConstraintParams* params, VecConstraintBlockInfo& blocks, VecPersistentID& ids, VecConstCoord& positions, VecConstDeriv& directions, VecConstArea& areas)
-        : simulation::BaseMechanicalVisitor(params)
-        , _blocks(blocks)
-        , _ids(ids)
-        , _positions(positions)
-        , _directions(directions)
-        , _areas(areas)
-        , _cparams(params)
-    {}
-
-    Result fwdConstraintSet(simulation::Node* node, core::behavior::BaseConstraintSet* cSet) override
-    {
-        if (core::behavior::BaseConstraint *c=cSet->toBaseConstraint())
-        {
-            ctime_t t0 = begin(node, c);
-            c->getConstraintInfo(_cparams, _blocks, _ids, _positions, _directions, _areas);
-            end(node, c, t0);
-        }
-        return RESULT_CONTINUE;
-    }
-
-
-    // This visitor must go through all mechanical mappings, even if isMechanical flag is disabled
-    bool stopAtMechanicalMapping(simulation::Node* /*node*/, core::BaseMapping* /*map*/) override
-    {
-        return false;
-    }
-
-    /// Return a class name for this visitor
-    /// Only used for debugging / profiling purposes
-    const char* getClassName() const override { return "MechanicalGetConstraintInfoVisitor";}
-
-private:
-    VecConstraintBlockInfo& _blocks;
-    VecPersistentID& _ids;
-    VecConstCoord& _positions;
-    VecConstDeriv& _directions;
-    VecConstArea& _areas;
-    const core::ConstraintParams* _cparams;
-};
 
 void LCPConstraintProblem::solveTimed(double tolerance, int maxIt, double timeout)
 {
